Extract BraveBatchPassageEmbedder to its own .h/.cc
Pure move — BraveBatchPassageEmbedder now owns factory init, the
receiver, and the renderer pipe for a single load, so it is a
standalone unit rather than a helper of BravePassageEmbeddingsService.
Splitting the files makes the two classes independently readable.
